+++ crates/client/rpc/src/methods/trace/trace_transaction.rs
@@ -17,7 +17,7 @@ use starknet_ff::FieldElement;
 use super::utils::{map_transaction_to_user_transaction, tx_execution_infos_to_tx_trace};
 use crate::errors::StarknetRpcApiError;
 use crate::madara_backend_client::get_block_by_block_hash;
-use crate::utils::utils::previous_substrate_block_hash;
+use crate::utils::helpers::previous_substrate_block_hash;
 use crate::Starknet;
 
 pub async fn trace_transaction<A, BE, G, C, P, H>(
