+++ crates/client/rpc/src/utils/mod.rs
@@ -1,4 +1,4 @@
 pub(crate) mod block;
 pub(crate) mod call_info;
+pub(crate) mod helpers;
 pub(crate) mod transaction;
-pub(crate) mod utils;
